Revert "test: introduce load test"
This reverts commit b3a29898fe42973b5ac5d7ec0860d4bc3d05b458.
