+++ README.md
@@ -50,10 +50,6 @@ Also check [AWS specifics configuration](https://github.com/elastic-ipfs/elastic
 
 **Note**: `DYNAMO_BLOCKS_TABLE` and `DYNAMO_CARS_TABLE` will be removed after the transition to the new database schema will be completed.
 
-## Test
-
-TODO e2e test instructions: scope, target, how to
-
 ## Issues
 
 Please report issues in the [elastic-ipfs/elastic-ipfs repo](https://github.com/elastic-ipfs/elastic-ipfs/issues).
