+++ package.json
@@ -9,7 +9,6 @@
     "test": "c8 --reporter=text --reporter=html tap --rcfile=test/utils/config.yml",
     "test:ci": "npm run lint && c8 --reporter=text --reporter=json --check-coverage --branches 90 --functions 90 --lines 90 --statements 90 tap --rcfile=test/utils/config.yml",
     "test:regression": "LOG_LEVEL=silent node test/e2e/regression.js",
-    "test:load": "LOG_LEVEL=silent node test/e2e/load.js",
     "format": "prettier -w src test",
     "lint": "eslint src test"
   },
@@ -38,7 +37,6 @@
     "xml-js": "^1.6.11"
   },
   "devDependencies": {
-    "@dnlup/doc": "^4.0.1",
     "@ipld/car": "^4.1.4",
     "@ipld/dag-pb": "^2.1.17",
     "autocannon": "^7.9.0",
