--- test/e2e/helper/chart/src.html
@@ -1,60 +1,0 @@
-<html>
-
-<head>
-  <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
-  <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
-  <script type="text/javascript">
-google.charts.load('current', {packages: ['corechart', 'line']});
-google.charts.setOnLoadCallback(drawLineColors);
-
-function drawLineColors() {
-      var dataCpu = new google.visualization.DataTable();
-      dataCpu.addColumn('number', 'X');
-      dataCpu.addColumn('number', 'CPU');
-
-      var dataMemory = new google.visualization.DataTable();
-      dataMemory.addColumn('number', 'X');
-      dataMemory.addColumn('number', 'Memory');
-
-      {data}
-
-      var optionsCpu = {
-        title: 'CPU',
-        hAxis: {
-          title: 'Time'
-        },
-        vAxis: {
-          title: 'CPU %'
-        },
-        colors: ['#a52714', '#097138']
-      };
-
-      var optionsMemory = {
-        title: 'Memory',
-        hAxis: {
-          title: 'Time'
-        },
-        vAxis: {
-          title: 'Memory (MB)'
-        },
-        colors: ['#a52714', '#097138']
-      };
-
-      var chartCpu = new google.visualization.LineChart(document.getElementById('chart_div_cpu'));
-      chartCpu.draw(dataCpu, optionsCpu);
-
-      var chartMemory = new google.visualization.LineChart(document.getElementById('chart_div_memory'));
-      chartMemory.draw(dataMemory, optionsMemory);
-    }
-
-  </script>
-</head>
-
-<body>
-  <pre>{avg}</pre>
-  <p>! note these stats include "http-proxy" and "autocannon" resources</p>
-  <div id="chart_div_cpu" style="width: 800px; height: 500px;"></div>
-  <div id="chart_div_memory" style="width: 800px; height: 500px;"></div>
-</body>
-
-</html>
