--- test/e2e/snaps/load/200-info-blocks.json
@@ -1,192 +1,0 @@
-{
-  "blocks": [
-    {
-      "cid": "QmUGsfJPhJ6CLuvnvcdvJH5jc7Yxf19pSD1jRn9wbsPCB1",
-      "type": "i"
-    },
-    {
-      "cid": "QmcBMGr3tUpKpDJMESH1JHPtWih2kjLRMRUZnBbeaW35Wn",
-      "type": "i"
-    },
-    {
-      "cid": "QmZvXJSNAcPzwQLHM3J7x2rWkGLgVrBBsDYKqF4pLo21Fs",
-      "type": "i"
-    },
-    {
-      "cid": "QmZxjT13tUa72FbaQg95pXG1h5t1vMPcBc4cYXjVKAsJxg",
-      "type": "i"
-    },
-    {
-      "cid": "QmRT1kpMn7ANggwsf31zVuXNUNwpHqt3u7DfKhEbtbftbM",
-      "type": "i"
-    },
-    {
-      "cid": "QmRa1K28hVCFZZZkSWyX8o3dbRjYZHcSVJaxQWURaaBG8z",
-      "type": "i"
-    },
-    {
-      "cid": "Qmecr5QoBcrWcj7xAwnkzBg4iQUW4ec81WewtvGhQyGYvW",
-      "type": "i"
-    },
-    {
-      "cid": "QmekymBc9FKr5DeVutHtDtqVCVTsCBWaZG4e7y6FcpxFKK",
-      "type": "i"
-    },
-    {
-      "cid": "QmdHPv94eJZQR4nTLyZXzc2kfmfVe29uA52ERYpyNchGqA",
-      "type": "i"
-    },
-    {
-      "cid": "Qmc88zbj9DYsnqqvgfw37HoKdc6kTgoigDnzmmjvbDg1Wc",
-      "type": "i"
-    },
-    {
-      "cid": "QmRdg2KxJUNyMsKy8oSD4FJxmfSZdsTXUBZjVt6wC2B7dR",
-      "type": "i"
-    },
-    {
-      "cid": "QmSU7dJZWkEkPgZRJa59nXkzXnrvBEh2SsLr3and3nCjcA",
-      "type": "i"
-    },
-    {
-      "cid": "QmbXUJfRx7pAuadgCBUHQ5bomxf5mYdXwaf14iksnvTCkP",
-      "type": "i"
-    },
-    {
-      "cid": "QmXHZrYrCou49HXqyPwpKxdB89vTxSdmZz1NyQtfZDzq85",
-      "type": "i"
-    },
-    {
-      "cid": "QmZJZyDqqY38kh58EzxWZr19taenXqHxa3Aq8WUzT1Jrxr",
-      "type": "i"
-    },
-    {
-      "cid": "QmejjL2H97S9gw8MZSWa5zyMMq4QrBmXsDAmjzt98b6egy",
-      "type": "i"
-    },
-    {
-      "cid": "QmSrPFjUbrwWiqgLArzn9Ey2vPjgU8vJLj5KPLKBa9CAqQ",
-      "type": "i"
-    },
-    {
-      "cid": "QmUsrHA3AkZBgXQ3C8ZJjQ1ypLpVHmWbjdLSHQmJYEbWXs",
-      "type": "i"
-    },
-    {
-      "cid": "Qmd5q8R9uoXZ4eRSWdidooexTMzjYjsz5noYKRL5yZCGqT",
-      "type": "i"
-    },
-    {
-      "cid": "QmYAv28ACuHxYKzu3LmzHp7LT5kDMkcuzC17CL26HTsRFB",
-      "type": "i"
-    },
-    {
-      "cid": "QmaLYGU9JxHSLbeq1am67ae582PHizgfpEouzpTnU7ZyR2",
-      "type": "i"
-    },
-    {
-      "cid": "QmaiZoBrCoMsWLYgwRn69zijcLChRAo88We42o7GiLj5Mc",
-      "type": "i"
-    },
-    {
-      "cid": "QmVTL5bCcNjoCb55MHvG7LQTx1E3a1fY4rRurVPRSwCDYd",
-      "type": "i"
-    },
-    {
-      "cid": "QmXyjXo93s41BLefqFFUfTVPyTCSrgYCdtdza5jxh7PH82",
-      "type": "i"
-    },
-    {
-      "cid": "QmUr57KZuwf9RdVPfMQkSquyr8W1UxQFDwg2M63yVb6gM7",
-      "type": "i"
-    },
-    {
-      "cid": "Qmc4ed4jPAkR9prkFA2JKo6xkPNiacSKN4D4jvC7yRy2sr",
-      "type": "i"
-    },
-    {
-      "cid": "QmT2x1Xqs2ohiY4em7DE8WGZDTPKBAtomoU3pR1ofPEjMt",
-      "type": "i"
-    },
-    {
-      "cid": "QmNWsGcgVNyzogVbXwxd6FRE1uWzri7sadbs6oL7DV67wd",
-      "type": "i"
-    },
-    {
-      "cid": "QmRMuBc9oRRcedCppGFPBE4k1yVZtNJW8MPgm99a9rJAwC",
-      "type": "i"
-    },
-    {
-      "cid": "Qmd2L2NScvBf5fUnTa9GVrqCHvdnxHbUhEAUpgyocnXthy",
-      "type": "i"
-    },
-    {
-      "cid": "QmZsXeQcqvrovZfmwhitpqXeufAopt8y71UwycXdz3jT3C",
-      "type": "i"
-    },
-    {
-      "cid": "QmfQyLR4CScmL1vkHMN1MGgHbFMSr3S6toszxdeaECmbPV",
-      "type": "i"
-    },
-    {
-      "cid": "QmQZHgbfpzDbQ58UiYhwCZvWpbP7xgeSAEAP61ExW4pbcV",
-      "type": "i"
-    },
-    {
-      "cid": "QmPY2kEtyH1uWXc131ALb9BzDuRrSiwKu3ASPsxfZGEgSf",
-      "type": "i"
-    },
-    {
-      "cid": "QmTPEraSW4AyVrpNYeEeD9kZ2jb6wBfbdBbQJC6zb84rzn",
-      "type": "i"
-    },
-    {
-      "cid": "QmbTLgB77wDu8pJTFJZmKsY2zVe9bYkRYpyf5fhRAVkReK",
-      "type": "i"
-    },
-    {
-      "cid": "QmW6HJS5bNZqLScJGdzj9Un7zJyFZwx8hRCvfCUpZnV7YG",
-      "type": "i"
-    },
-    {
-      "cid": "QmU9d2EMV5GpF58XDiSKFKKLAp45C51dyGBW94FPAVCkJG",
-      "type": "i"
-    },
-    {
-      "cid": "QmbyVEzrQrK2Yy3SEDcPzyw8hL5v55PXggow8EGveoCZmi",
-      "type": "i"
-    },
-    {
-      "cid": "QmTfs36Yw9NuDPZU15axN1Gve7UPURjySCWFL1NSiR29Wr",
-      "type": "i"
-    },
-    {
-      "cid": "QmV9V6yJVvWjgzbYdLxJQdxLcjk3HFGTWGRRXfvQy1KvBq",
-      "type": "i"
-    },
-    {
-      "cid": "QmRr2daSAtRXZPfFbUBpriSGR6z7zDfgNgnJ9njj5yhsDp",
-      "type": "i"
-    },
-    {
-      "cid": "QmQnu9EoT57X5NAskDNGChdL7akWSdPJ5TTJXmH1YozU5o",
-      "type": "i"
-    },
-    {
-      "cid": "QmfCTPtUM9esvms5bmwRcaqeGrf2WzSeRaHdFXehQyuppP",
-      "type": "i"
-    },
-    {
-      "cid": "QmWkAcFu38kSs6VuY1VuYQTfQf1DqJgX8Jz8Nw7F6Xu4Vf",
-      "type": "i"
-    },
-    {
-      "cid": "QmcdTVRj3kqtTK9HNrRHVeGQsf5jfX7W9P2SZpmfRY59EF",
-      "type": "i"
-    },
-    {
-      "cid": "QmY1y2ZszqrHrKKWdYNw1PfQnVa25avJmjjsEry41p3pqz",
-      "type": "i"
-    }
-  ]
-}
